Split header nav helpers
